set a check for placing a bet on a game that's over, cannot place a bet on a live or completed game now. created a separate file for the odds clean up function
